Use objects to hold data and references to controls in order to simplify parameter passing, move processes to standalone file (closes #18)
