+++ semestralka-routing-simulation/semestralka-routing-simulation/Simulation.cs
@@ -38,419 +38,6 @@ namespace semestralka_routing_simulation
         }
     }
 
-    // Links are one-directional, i.e. for two connected devices there are two links.
-    // The network itself isn't directed, but this representation is more practical.
-    class Link : Process
-    {
-        public int toID, timeToTransfer;
-        public bool busy;
-        public Device sourceDevice;
-
-        public Link(int ID, int toID, int timeToTransfer, Device sourceDevice)
-        {
-            this.ID = ID;
-            this.toID = toID;
-            this.timeToTransfer = timeToTransfer;
-            this.sourceDevice = sourceDevice;
-            busy = false;
-            packetsOut = new List<Packet>();
-            packetsOutTimeouts = new List<ulong>();
-        }
-
-        public override void HandleEvent(SimulationEvent simEvent, Model model)
-        {
-            if (simEvent.eventType == EventType.SendPacket)
-            {
-                if (packetsOut.Count > 0 && !busy)
-                {
-                    Packet packet = packetsOut[0];
-                    ulong packetTimeout = packetsOutTimeouts[0];
-                    
-                    if (model.time <= packet.timeout && packetTimeout == packet.timeout)
-                    {
-                        SimulationEvent finishSending = new SimulationEvent(model.time + (ulong)timeToTransfer, this, EventType.FinishSending);
-                        model.scheduler.Add(finishSending);
-                        busy = true;
-                    }
-                    else
-                    {
-                        packetsOut.RemoveAt(0);
-                        packetsOutTimeouts.RemoveAt(0);
-                        SimulationEvent trySendAnother = new SimulationEvent(model.time, this, EventType.SendPacket);
-                        model.scheduler.Add(trySendAnother);
-
-                        Debug.WriteLine($"Packet {packet.ID} timed out");
-                    }
-                }
-            }
-            else if (simEvent.eventType == EventType.FinishSending)
-            {
-                Packet packet = packetsOut[0];
-                packetsOut.RemoveAt(0);
-
-                ulong packetTimeout = packetsOutTimeouts[0];
-                packetsOutTimeouts.RemoveAt(0);
-
-                if (model.time <= packet.timeout && packetTimeout == packet.timeout)
-                {
-                    int destinationRoutingIndex = Simulation.deviceIndexToRoutingIndex[packet.destination];
-                    int nextHopRoutingIndex = model.routingTable[sourceDevice.routingTableIndex, destinationRoutingIndex];
-                    Device nextHopDevice = Simulation.routingIndexToDevice[nextHopRoutingIndex];
-
-                    SimulationEvent receiveEvent = new SimulationEvent(model.time, nextHopDevice, EventType.ProcessPacket);
-                    model.scheduler.Add(receiveEvent);
-                    nextHopDevice.addPacketIn(packet);
-                    nextHopDevice.addPacketInTimeout(packet.timeout);
-                }
-                else
-                {
-                    Debug.WriteLine($"Packet {packet.ID} timed out");
-                }
-
-                SimulationEvent trySendAnother = new SimulationEvent(model.time, this, EventType.SendPacket);
-                model.scheduler.Add(trySendAnother);
-                busy = false;
-            }
-        }
-    }
-
-    abstract class Process
-    {
-        public int ID;
-        protected List<Packet> packetsOut;
-        protected List<ulong> packetsOutTimeouts;
-
-        public abstract void HandleEvent(SimulationEvent simEvent, Model model);
-        public void addPacketOut(Packet packetOut)
-        {
-            packetsOut.Add(packetOut);
-        }
-        public void addPacketOutTimeout(ulong timeout)
-        {
-            packetsOutTimeouts.Add(timeout);
-        }
-    }
-
-    abstract class Device : Process
-    {
-        public int routingTableIndex;
-        protected List<Link> links;
-        protected List<Packet> packetsIn;
-        protected List<ulong> packetsInTimeouts;
-
-        public void addLink(Link link)
-        {
-            links.Add(link);
-        }
-        public void addPacketIn(Packet packetIn)
-        {
-            packetsIn.Add(packetIn);
-        }
-        public void addPacketInTimeout(ulong timeout)
-        {
-            packetsInTimeouts.Add(timeout);
-        }
-        protected Link getLink(Packet packet, Model  model)
-        {
-            int destinationRoutingIndex = Simulation.deviceIndexToRoutingIndex[packet.destination];
-            int nextHopRoutingIndex = model.routingTable[routingTableIndex, destinationRoutingIndex];
-            Device nextHopDevice = Simulation.routingIndexToDevice[nextHopRoutingIndex];
-
-            Link link = null;
-            foreach (Link l in links)
-            {
-                if (l.toID == nextHopDevice.ID)
-                {
-                    link = l;
-                }
-            }
-
-            return link;
-        }
-    }
-    
-    class Router : Device
-    {
-        public int timeToProcess;
-        public bool processing;
-        Firewall firewall;
-        public Router(int id, int timeToProcess, int routingTableIndex)
-        {
-            this.ID = id;
-            this.timeToProcess = timeToProcess;
-            this.routingTableIndex = routingTableIndex;
-            packetsIn = new List<Packet>();
-            packetsOut = new List<Packet>();
-            links = new List<Link>();
-            packetsOutTimeouts = new List<ulong>();
-            packetsInTimeouts = new List<ulong>();
-            processing = false;
-            firewall = null;
-        }
-
-        public void setFirewall(Firewall firewall)
-        {
-            this.firewall = firewall;
-        }
-        
-        public override void HandleEvent(SimulationEvent simEvent, Model model)
-        {
-            if (simEvent.eventType == EventType.SendPacket)
-            {
-                Packet packet = packetsOut[0];
-                packetsOut.RemoveAt(0);
-
-                ulong packetTimeout = packetsOutTimeouts[0];
-                packetsOutTimeouts.RemoveAt(0);
-
-                Link link = getLink(packet, model);
-
-                SimulationEvent sendPacket = new SimulationEvent(model.time, link, EventType.SendPacket);
-                link.addPacketOut(packet);
-                link.addPacketOutTimeout(packetTimeout);
-                model.scheduler.Add(sendPacket);
-            }
-            else if (simEvent.eventType == EventType.ProcessPacket)
-            {
-                if (packetsIn.Count > 0 && !processing)
-                {
-                    Packet packet = packetsIn[0];
-                    ulong packetTimeout = packetsInTimeouts[0];
-
-                    if (model.time <= packet.timeout && packetTimeout == packet.timeout)
-                    {
-                        processing = true;
-                        SimulationEvent processingFinished = new SimulationEvent(model.time + (ulong)timeToProcess, this, EventType.FinishProcessing);
-                        model.scheduler.Add(processingFinished);
-                    }
-                    else
-                    {
-                        packetsIn.RemoveAt(0);
-                        packetsInTimeouts.RemoveAt(0);
-                        SimulationEvent tryProcessAnother = new SimulationEvent(model.time, this, EventType.ProcessPacket);
-                        model.scheduler.Add(tryProcessAnother);
-
-                        Debug.WriteLine($"Packet {packet.ID} timed out");
-                    }
-                }
-            }
-            else if (simEvent.eventType == EventType.FinishProcessing)
-            {
-                processing = false;
-                Packet packet = packetsIn[0];
-                packetsIn.RemoveAt(0);
-
-                ulong packetTimeout = packetsInTimeouts[0];
-                packetsInTimeouts.RemoveAt(0);
-
-                if (model.time <= packet.timeout && packetTimeout == packet.timeout)
-                {
-                    if (firewall != null)
-                    {
-                        SimulationEvent processPacket = new SimulationEvent(model.time, firewall, EventType.ProcessPacket);
-                        firewall.addPacketOut(packet);
-                        firewall.addPacketOutTimeout(packetTimeout);
-                        model.scheduler.Add(processPacket);
-                    }
-                    else
-                    {
-                        SimulationEvent sendPacket = new SimulationEvent(model.time, this, EventType.SendPacket);
-                        addPacketOut(packet);
-                        addPacketOutTimeout(packetTimeout);
-                        model.scheduler.Add(sendPacket);
-                    }
-                }
-                else
-                {
-                    Debug.WriteLine($"Packet {packet.ID} timed out");
-                }
-
-                SimulationEvent tryProcessAnother = new SimulationEvent(model.time, this, EventType.ProcessPacket);
-                model.scheduler.Add(tryProcessAnother);
-            }
-        }
-    }
-    
-    class Computer : Device
-    {
-
-        List<Packet> packetsSent;
-        
-        public Computer(int id, int routingTableIndex)
-        {
-            this.ID = id;
-            this.routingTableIndex = routingTableIndex;
-            packetsIn = new List<Packet>();
-            packetsOut = new List<Packet>();
-            packetsSent = new List<Packet>();
-            links = new List<Link>();
-            packetsOutTimeouts = new List<ulong>();
-            packetsInTimeouts = new List<ulong>();
-        }
-        public override void HandleEvent(SimulationEvent simEvent, Model model)
-        {
-            if (simEvent.eventType == EventType.SendPacket)
-            {
-                Packet packet = packetsOut[0];
-                packetsOut.RemoveAt(0);
-
-                packet.timeFirstSent = model.time;
-                
-                sendPacket(packet, model);
-
-                if (packet.malicious)
-                {
-                    model.statistics.sentPacketsMalicious += 1;
-                } 
-                else
-                {
-                    model.statistics.sentPackets += 1;
-                }
-
-                Debug.WriteLine($"{model.time}: Computer {ID} sends packet {packet.ID}");
-            }
-            else if (simEvent.eventType == EventType.ProcessPacket)
-            {
-                Packet packet = packetsIn[0];
-                packetsIn.RemoveAt(0);
-
-                ulong packetTimeout = packetsInTimeouts[0];
-                packetsInTimeouts.RemoveAt(0);
-                
-                if (model.time <= packet.timeout && packetTimeout == packet.timeout)
-                {
-                    packet.received = true;
-
-                    model.statistics.deliveredPackets += 1;
-                    model.statistics.totalDeliveryTime += model.time - packet.timeFirstSent;
-                    if (packet.malicious)
-                    {
-                        model.statistics.deliveredPacketsMalicious += 1;
-                    }
-                    else
-                    {
-                        model.statistics.totalNumberAttempts += packet.attemptNumber;
-                    }
-
-                    Debug.WriteLine($"Computer {ID} received packet with id {packet.ID} from computer {packet.source} in time {model.time}, malicious: {packet.malicious}, attempt number: {packet.attemptNumber}");
-                }
-                else
-                {
-                    Debug.WriteLine($"Packet {packet.ID} timed out");
-                }
-                
-            }
-            else if (simEvent.eventType == EventType.Timeout)
-            {
-                // Packets timeout in the order they are sent
-                Packet packet = packetsSent[0];
-                packetsSent.RemoveAt(0);
-
-                if (!packet.received)
-                {
-                    if (packet.attemptNumber < model.timeoutAttempts)
-                    {
-                        sendPacket(packet, model);
-                        packet.attemptNumber += 1;
-                    }
-                    else
-                    {
-                        Debug.WriteLine($"Packet {packet.ID} from computer {ID} couldn't be delivered in {model.timeoutAttempts} attempts");
-                    }
-                }
-
-            }
-        }
-        private void sendPacket(Packet packet, Model model)
-        {
-            packet.timeout = model.time + model.timeout;
-
-            Link link = getLink(packet, model);
-
-            SimulationEvent sendPacket = new SimulationEvent(model.time, link, EventType.SendPacket);
-            link.addPacketOut(packet);
-            link.addPacketOutTimeout(packet.timeout);
-            model.scheduler.Add(sendPacket);
-
-            // Adding a single tick because timeout is inclusive (i.e. if packet arrives precisely in time of timeout,
-            // it is still considered to be received and it shouldn't be resent)
-            SimulationEvent resolveTimeout = new SimulationEvent(packet.timeout + 1, this, EventType.Timeout);
-            packetsSent.Add(packet);
-            model.scheduler.Add(resolveTimeout);
-        }
-    }
-    
-    class Firewall : Process
-    {
-        public int timeToProcess;
-        public bool processing;
-        Router router;
-        public Firewall(int id, int timeToProcess, Router router)
-        {
-            this.ID = id;
-            this.timeToProcess = timeToProcess;
-            packetsOut = new List<Packet>();
-            packetsOutTimeouts = new List<ulong>();
-            processing = false;
-            this.router = router;
-        }
-        public override void HandleEvent(SimulationEvent simEvent, Model model)
-        {
-            if (simEvent.eventType == EventType.ProcessPacket)
-            {
-                if (packetsOut.Count > 0 && !processing)
-                {
-                    Packet packet = packetsOut[0];
-                    ulong packetTimeout = packetsOutTimeouts[0];
-
-                    if (model.time <= packet.timeout && packetTimeout == packet.timeout)
-                    {
-                        processing = true;
-                        SimulationEvent processingFinished = new SimulationEvent(model.time + (ulong)timeToProcess, this, EventType.FinishProcessing);
-                        model.scheduler.Add(processingFinished);
-                    }
-                    else
-                    {
-                        packetsOut.RemoveAt(0);
-                        packetsOutTimeouts.RemoveAt(0);
-                        SimulationEvent tryProcessAnother = new SimulationEvent(model.time, this, EventType.ProcessPacket);
-                        model.scheduler.Add(tryProcessAnother);
-
-                        Debug.WriteLine($"Packet {packet.ID} timed out");
-                    }
-                }
-            }
-            else if (simEvent.eventType == EventType.FinishProcessing)
-            {
-                processing = false;
-                Packet packet = packetsOut[0];
-                packetsOut.RemoveAt(0);
-
-                ulong packetTimeout = packetsOutTimeouts[0];
-                packetsOutTimeouts.RemoveAt(0);
-
-                if (packet.malicious)
-                {
-                    Debug.WriteLine($"Firewall {this.ID} found out that packet {packet.ID} is malicious and discarded it.");
-                }
-                else if (model.time > packet.timeout || packetTimeout != packet.timeout)
-                {
-                    Debug.WriteLine($"Packet {packet.ID} timed out");
-                }
-                else
-                {
-                    SimulationEvent sendPacket = new SimulationEvent(model.time, router, EventType.SendPacket);
-                    router.addPacketOut(packet);
-                    router.addPacketOutTimeout(packetTimeout);
-                    model.scheduler.Add(sendPacket);
-                }
-
-                SimulationEvent tryProcessAnother = new SimulationEvent(model.time, this, EventType.ProcessPacket);
-                model.scheduler.Add(tryProcessAnother);
-            }
-        }
-    }
-
     class SimulationEvent 
     {
         public ulong time;
@@ -541,9 +128,9 @@ namespace semestralka_routing_simulation
         public static Dictionary<int, int> deviceIndexToRoutingIndex = new Dictionary<int, int>();
         public static Dictionary<int, int> routingIndexToDeviceIndex = new Dictionary<int, int>();
 
-        public static void extractDevices(Model model, ListBox devices)
+        public static void extractDevices(Model model, ListBox.ObjectCollection devices)
         {
-            foreach (FormDevice device in devices.Items)
+            foreach (FormDevice device in devices)
             {
                 if (device.deviceType == DeviceType.Computer)
                 {
@@ -595,14 +182,14 @@ namespace semestralka_routing_simulation
             return routing_table;
         }
 
-        public static void initializeRoutingTables(Model model, ListBox devices)
+        public static void initializeRoutingTables(Model model, ListBox.ObjectCollection devices)
         {
-            int[,] routingTableSuccesors = new int[devices.Items.Count, devices.Items.Count];
-            ulong[,] routingTableWeights = new ulong[devices.Items.Count, devices.Items.Count];
+            int[,] routingTableSuccesors = new int[devices.Count, devices.Count];
+            ulong[,] routingTableWeights = new ulong[devices.Count, devices.Count];
             
-            for (int i = 0; i < devices.Items.Count; i++)
+            for (int i = 0; i < devices.Count; i++)
             {
-                for (int j = 0; j < devices.Items.Count; j++)
+                for (int j = 0; j < devices.Count; j++)
                 {
                     if (i == j)
                     {
@@ -619,7 +206,7 @@ namespace semestralka_routing_simulation
                 }
             }
 
-            foreach (FormDevice device in devices.Items)
+            foreach (FormDevice device in devices)
             {
                 int sourceIndex = deviceIndexToRoutingIndex[device.ID];
 
@@ -636,9 +223,9 @@ namespace semestralka_routing_simulation
             model.routingTableWeights = routingTableWeights;
         }
 
-        public static void assignRoutingIndices(ListBox devices)
+        public static void assignRoutingIndices(ListBox.ObjectCollection devices)
         {
-            foreach (FormDevice device in devices.Items)
+            foreach (FormDevice device in devices)
             {
                 deviceIndexToRoutingIndex.Add(device.ID, routingIndex);
                 routingIndexToDeviceIndex.Add(routingIndex, device.ID);
@@ -646,16 +233,15 @@ namespace semestralka_routing_simulation
             }
         }
 
-        public static void updateResults(Statistics statistics, TextBox simulationLength, TextBox packetsSent, TextBox packetsDelivered, TextBox packetsSentMalicious, 
-            TextBox packetsDeliveredMalicious, TextBox averageTimeDelivered, TextBox averageAttempts)
+        public static void updateResults(Statistics statistics, ResultControls controls)
         {
-            changeText(simulationLength, statistics.lengthOfSimulation.ToString());
-            changeText(packetsSent, statistics.sentPackets.ToString());
-            changeText(packetsDelivered, statistics.deliveredPackets.ToString());
-            changeText(packetsSentMalicious, statistics.sentPacketsMalicious.ToString());
-            changeText(packetsDeliveredMalicious, statistics.deliveredPacketsMalicious.ToString());
-            changeText(averageTimeDelivered, statistics.getAverageDeliveryTime().ToString());
-            changeText(averageAttempts, statistics.getAverageNumberAttempts().ToString());
+            changeText(controls.simulationLength, statistics.lengthOfSimulation.ToString());
+            changeText(controls.packetsSent, statistics.sentPackets.ToString());
+            changeText(controls.packetsDelivered, statistics.deliveredPackets.ToString());
+            changeText(controls.packetsSentMalicious, statistics.sentPacketsMalicious.ToString());
+            changeText(controls.packetsDeliveredMalicious, statistics.deliveredPacketsMalicious.ToString());
+            changeText(controls.averageTimeDelivered, statistics.getAverageDeliveryTime().ToString());
+            changeText(controls.averageAttempts, statistics.getAverageNumberAttempts().ToString());
         }
 
         public static void changeText(TextBox output, string text)
@@ -682,6 +268,16 @@ namespace semestralka_routing_simulation
             {
                 panelInput.Enabled = true;
             }
+        }
+
+        public static void resetSimulationParameters()
+        {
+            routingIndex = 0;
+            linkIndex = 1;
+            firewallIndex = 1;
+            routingIndexToDevice = new Dictionary<int, Device>();
+            deviceIndexToRoutingIndex = new Dictionary<int, int>();
+            routingIndexToDeviceIndex = new Dictionary<int, int>();
         }
 
         // Normal distribution with mean = maxTime / 2 and std = maxTime / 4
@@ -749,19 +345,16 @@ namespace semestralka_routing_simulation
             return BitConverter.ToUInt64(buffer, 0) % (maxValue + 1);
         }
 
-        public static void RunSimulation(ListBox devices, NumericUpDown totalPackets, NumericUpDown maxTime, RadioButton distributionUniform, 
-            NumericUpDown probabilityMalicious, NumericUpDown timeoutAttempts, NumericUpDown timeout, NumericUpDown randomSeed,
-            TextBox simulationLength, TextBox packetsSent, TextBox packetsDelivered, TextBox packetsSentMalicious,
-            TextBox packetsDeliveredMalicious, TextBox averageTimeDelivered, TextBox averageAttempts, Panel panelInput)
+        public static void RunSimulation(SimulationParametersDto simulationParameters, ResultControls controls, Panel panelInput)
         {
             Debug.WriteLine("Simulation started");
 
             Scheduler scheduler = new Scheduler();
             Statistics statistics = new Statistics();
-            Model model = new Model(scheduler, (ulong)timeout.Value, (int)timeoutAttempts.Value, statistics);
+            Model model = new Model(scheduler, simulationParameters.timeout, simulationParameters.numberAttempts, statistics);
 
             Distribution distribution;
-            if (distributionUniform.Checked)
+            if (simulationParameters.distributionUniform)
             {
                 distribution = Distribution.Uniform;
             }
@@ -770,11 +363,11 @@ namespace semestralka_routing_simulation
                 distribution = Distribution.DiscreteGaussian;
             }
 
-            assignRoutingIndices(devices);
-            extractDevices(model, devices);
-            initializeRoutingTables(model, devices);
-            generatePackets(model.computers, scheduler, (ulong)totalPackets.Value, (ulong)maxTime.Value, (double)probabilityMalicious.Value, 
-                statistics, distribution, (int)randomSeed.Value);
+            assignRoutingIndices(simulationParameters.devices);
+            extractDevices(model, simulationParameters.devices);
+            initializeRoutingTables(model, simulationParameters.devices);
+            generatePackets(model.computers, scheduler, simulationParameters.totalPackets, simulationParameters.sendUntil, simulationParameters.probabilityMalicious, 
+                statistics, distribution, simulationParameters.randomSeed);
 
             // TODO remove this line when Floyd-Warshall is implemented
             model.routingTable = getRouting();
@@ -789,8 +382,7 @@ namespace semestralka_routing_simulation
                 if (model.time != lastTime)
                 {
                     statistics.lengthOfSimulation = model.time;
-                    updateResults(statistics, simulationLength, packetsSent, packetsDelivered, packetsSentMalicious, packetsDeliveredMalicious, 
-                        averageTimeDelivered, averageAttempts);
+                    updateResults(statistics, controls);
                     lastTime = model.time;
                 }
             }
@@ -801,13 +393,7 @@ namespace semestralka_routing_simulation
             Debug.WriteLine($"Average time of delivery: {statistics.getAverageDeliveryTime()}");
             Debug.WriteLine($"Average number of attempts: {statistics.getAverageNumberAttempts()}");
 
-            routingIndex = 0;
-            linkIndex = 1;
-            firewallIndex = 1;
-            routingIndexToDevice = new Dictionary<int, Device>();
-            deviceIndexToRoutingIndex = new Dictionary<int, int>();
-            routingIndexToDeviceIndex = new Dictionary<int, int>();
-
+            resetSimulationParameters();
             enableInput(panelInput);
         }
     }
